# Pass affected labels to `MemPostings.Delete()`