As [suggested](https://github.com/prometheus/prometheus/pull/13286#pullrequestreview-2107480790) by @bboreham, we can track the labels of the deleted series and avoid iterating through all the label/value combinations.

This looks much faster on the MemPostings.Delete call. We don't have a benchmark on stripeSeries.gc() where we'll pay the price of iterating the labels of each one of the deleted series.

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb/index
                                            │      main      │              affected              │
                                            │     sec/op     │   sec/op     vs base               │
MemPostings_Delete/refs=1/readers=0-12         139.53m ±  4%   23.81m ± 3%  -82.94% (p=0.002 n=6)
MemPostings_Delete/refs=1/readers=1-12         176.45m ±  1%   23.66m ± 2%  -86.59% (p=0.002 n=6)
MemPostings_Delete/refs=1/readers=10-12       1171.34m ± 48%   23.95m ± 1%  -97.96% (p=0.002 n=6)
MemPostings_Delete/refs=100/readers=0-12       171.41m ± 11%   34.84m ± 5%  -79.67% (p=0.002 n=6)
MemPostings_Delete/refs=100/readers=1-12       213.05m ± 11%   35.94m ± 4%  -83.13% (p=0.002 n=6)
MemPostings_Delete/refs=100/readers=10-12      863.45m ± 36%   36.39m ± 4%  -95.79% (p=0.002 n=6)
MemPostings_Delete/refs=10000/readers=0-12     168.06m ±  6%   37.12m ± 3%  -77.91% (p=0.002 n=6)
MemPostings_Delete/refs=10000/readers=1-12     225.20m ±  3%   37.26m ± 1%  -83.46% (p=0.002 n=6)
MemPostings_Delete/refs=10000/readers=10-12   1019.95m ± 48%   37.78m ± 8%  -96.30% (p=0.002 n=6)
geomean                                         319.9m         31.68m       -90.10%
```