@colega 2024-06-17
I have added a benchmark for `Head.Truncate()`, which calls `head.gc()` so it benchmarks the combination of `MemPostings.Delete()` and `stripeSeries.gc()`, the results look good:

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
                            │     main     │               affected               │
                            │    sec/op    │    sec/op     vs base                │
Head_Truncate/churn=10-12     319.4m ± 16%   125.1m ± 10%  -60.81% (p=0.000 n=10)
Head_Truncate/churn=100-12    305.2m ±  3%   138.9m ± 12%  -54.50% (p=0.000 n=10)
Head_Truncate/churn=1000-12   298.2m ±  4%   136.1m ± 21%  -54.36% (p=0.000 n=10)
geomean                       307.5m         133.2m        -56.66%

                            │     main     │              affected               │
                            │     B/op     │     B/op      vs base               │
Head_Truncate/churn=10-12     28.63Mi ± 0%   28.48Mi ± 0%  -0.54% (p=0.000 n=10)
Head_Truncate/churn=100-12    29.04Mi ± 0%   28.92Mi ± 0%  -0.42% (p=0.000 n=10)
Head_Truncate/churn=1000-12   29.36Mi ± 0%   29.48Mi ± 0%  +0.42% (p=0.000 n=10)
geomean                       29.01Mi        28.96Mi       -0.18%

                            │    main     │              affected              │
                            │  allocs/op  │  allocs/op   vs base               │
Head_Truncate/churn=10-12     14.98k ± 0%   14.98k ± 0%  -0.01% (p=0.000 n=10)
Head_Truncate/churn=100-12    15.24k ± 0%   15.25k ± 0%  +0.07% (p=0.000 n=10)
Head_Truncate/churn=1000-12   17.68k ± 0%   17.71k ± 0%  +0.17% (p=0.000 n=10)
geomean                       15.92k        15.93k       +0.08%
```